requirements.txt
bcrypt>=4.0
pywin32>=305
pystray>=0.19
Pillow>=10.0
pyinstaller>=6.0
pytest>=8.0
